Move url selectors to their own subfolder
