+++ src/Main/Layout/NavigationBar.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 
-import { getFightId, getPlayerName } from 'selectors/routing';
+import { getFightId, getPlayerName } from 'selectors/url';
 import { getReport } from 'selectors/report';
 import { getFightById } from 'selectors/fight';
