+++ src/selectors/url/getPlayerName.js
@@ -1,0 +1,6 @@
+import getMatch from './getMatch';
+
+export default state => {
+  const match = getMatch(state);
+  return match ? match.params.playerName : null;
+};
