+++ src/selectors/url/getResultTab.js
@@ -1,0 +1,6 @@
+import getMatch from './getMatch';
+
+export default state => {
+  const match = getMatch(state);
+  return match ? match.params.resultTab : null;
+};
