+++ src/selectors/url/index.js
@@ -1,0 +1,4 @@
+export { default as getReportCode } from './getReportCode';
+export { default as getFightId } from './getFightId';
+export { default as getPlayerName } from './getPlayerName';
+export { default as getResultTab } from './getResultTab';
